+++ src/constants/Types.ts
@@ -3,7 +3,7 @@ import React from "react";
 // Interfaces
 export interface UserRankingPropTypes {
     currentPoints: number;
-    username: string | TelegramUserData;
+    username: string;
 }
 
 export interface EnergyMeterPropTypes {
@@ -22,6 +22,12 @@ export interface TelegramUserData {
     is_premium: boolean;
 }
 
+export interface DashboardProps {
+    currentPoints: number;
+    energy: number;
+    setCurrentPoints: React.Dispatch<React.SetStateAction<number>>,
+    setEnergy: React.Dispatch<React.SetStateAction<number>>,
+}
 
 // Types
 export type clickType = {